Split delegation signer record specs from dnssec specs.
